requirements/common.txt
# requirements/common.txt
Django
South
django-admin-bootstrapped
django-autoslug
django-tinymce
pillow
docutils
python-magic
MySQL-python
django-debug-toolbar
django-nose
djangorestframework
markdown
django-filter
# Add on to make the DRF structure what EmberJS wants.
rest-framework-ember
django-cors-headers
sorl-thumbnail
humanize